Move MainActivity into presentation.main package
