+++ app/src/main/java/com/yang/a09p9/presentation/main/MainFragment.kt
@@ -1,5 +1,6 @@
-package com.yang.a09p9
+package com.yang.a09p9.presentation.main
 
+import com.yang.a09p9.R
 import com.yang.a09p9.base.BaseFragment
 import com.yang.a09p9.databinding.FragmentMainBinding
