+++ app/src/main/res/layout/activity_main.xml
@@ -6,13 +6,13 @@
     <data>
         <variable
             name="activity"
-            type="com.yang.a09p9.presentation.MainActivity" />
+            type="com.yang.a09p9.presentation.main.MainActivity" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.MainActivity">
+        tools:context=".presentation.main.MainActivity">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/mainFragmentContainerView"
