+++ app/src/main/res/layout/fragment_main.xml
@@ -9,7 +9,8 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainFragment">
+        tools:context=".presentation.main.MainFragment">
+
 
 
     </androidx.constraintlayout.widget.ConstraintLayout>
