+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="com.yang.a09p9.MainFragment"
+        android:name="com.yang.a09p9.presentation.main.MainFragment"
         android:label="fragment_main"
         tools:layout="@layout/fragment_main" />
 </navigation>
